[FIRRTL] Fix post order traversal in RemoveUnusedPorts
The `RemoveUnusedPortsPass` intends to do a traversal of the module
hierarchy from leaves to the top, but it uses the basic node iterator
of the `instanceGraph`, which doesn't give any traversal guarantees. (I
think it's just using the declaration order in the input.)

This commit fixes the pass to use `llvm::post_order` to enforce a post
order traversal over the modules, as the pass intends to do per its
comments. This will probably pick up additional optimization
opportunities that were missed before due to the suboptimal traversal
order before.