[FIRRTL] Add RemoveUnusedPorts option to ignore dont-touch/annos
The `RemoveUnusedPortsPass` currently does not remove ports if they are
marked as don't-touch, i.e., they have a symbol, or if they have any
annotations on them.

This commit adds a `ignoreDontTouch` option to the pass to cause it to
still remove ports even if they are marked as described above. This is
useful for example in the reducer, where preserving the functional
correctness of the input is not necessarily required, and more
aggressive removal is desired.