[FIRRTL][Dedup] Include port types in ext module hashing (#2721)
When hashing port types, the hasher was ignoring the `portType`
attribute of the operation, and it would hash the type of each block
argument.  In other words, it didn't make sense to hash the type twice.

This was a huge problem for extmodules, which do not have block
arguments. This changes the hasher to use the `portTypes` attribute
instead of hashing the block argument types.