[FIRRTL][OMIR] Don't assume NLA's have an absolute path (#2747)
* Don't assume NLA's have an absolute path

NLAs are used by EmitOMIR to print the instance path to memories and
modules. When an OMIR annotation does not have an NLA, a new absolute
NLA is created for it.  As a part of this, the pass assumes that NLAs
used by any OMIR node is rooted at the top module.  Dedup notably will
create NLAs for OMIR annotations which are not rooted at the root
module, and as a result the final OMIR JSON file will include only part
of the path to the memories or instances.

This changes the metadata emission to stop assuming that NLAs are rooted
at the top module, and instead use the existing NLA as a way to
disambiguate the path to the memory when creating a new absolute NLA.
The pass already will remove any "temporary" NLAs created for metadata
emission.

* Don't add tracker to leaf

* Don't change pass behavior for module due to one memory

Co-authored-by: Andrew Lenharth <andrew@lenharth.org>