[FIRRTL][Inliner] Filter non-local annos by the instance path (#2770)
When inlining a module, we copy all the operations with their
annotations into another module.  When the copied operations have
non-local annotations attached to them, it doesn't make sense to copy
the annotation when the NLA's namepath does not match match the current
module path, and the annotation could never be considered active.

In addition to the above, inlining can make a non-local annotation turn
in to a local annotation as it inlines modules.  This change fixes a bug
where we copy an annotation which should never be active according to
the instance path, and simulatiously change it to a local annotation,
causing the annotation to proliferate to all places the operation is
cloned.

To fix this, the change keeps a stack of the current instance path as it
recursively inlines modules. When the current instance path does not
match the path in an NLA, the corresponding annotation is not copied.