[Moore] Add simple bit vector type
Add the `SimpleBitVectorType` helper struct. SystemVerilog uses these
types throughout the standard to deal with various packed types and how
they cast amongst each other. For example, `bit`, `logic [0:0]`,
`reg [31:0]`, or `int` are SBVs, but `bit [1:0][2:0]`, `int [4:0]`,
`bit [5:2]`, or `bit []` are not.

This new struct is not a type in its own right, but offers a condensed
nugget of information that fully captures and SBV. Other Moore types can
be cast to a `SimpleBitVectorType`, and that in turn can be converted
back to a regular Moore type. A lot of the casting logic in SV relies on
being able to go from any type to the corresponding SBV type, perform
the casting (e.g. altering the sign or size), and then going back to the
"most appropriate" type again. This new struct allows for that to
happen.