[FIRRTL] Do not add debug info in BlackBoxReader
Change BlackBoxReader to not add location debug information to verbatim
statements that hold black box contents.  This has the effect of
preventing this debug information from showing up in the output file.
This is done because the API around black boxes is that they are copied
exactly and not modified.  (This avoids a hypothetical situation where
the comment syntax for ExportVerilog, "//", may not be a comment in the
type of file which is the black box---blackboxes can be any file type,
not just Verilog.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>