[ExportVerilog] Fix Mul/Add Width Lint Errors
Change emission of hw.mul and hw.add to pattern match the situation
where these are zero-padded to align with HW dialect width rules.  For
situations where we can drop these pads, do it.  This is done to work
around lint rules that are upset if you truncate the result of one of
these large multiplication/additions.

E.g., previously we would emit code like:

    wire [4:0] a, b;
    wire [9:0] c = {5'0, a} * {5'0, b};

Now we will always emit:

    wire [4:0] a, b;
    wire [9:0] c = a * b;

This is a replacement, default emission strategy for the
"explicitBitcastAddMul" lowering option which does not work in all
cases, e.g., for situations involving muxing true/false legs that are
multiplications or additions.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>