[FIRRTL][Dedup] Don't dump the circuit on errors (#2834)
The "MustDedup" annotation causes some user-directed error checking, and
I decided to use the circuit op for the error messages.  I think that
this was a mistake; this causes the entire circuit operation to be
dumped.  This change it to print the error message using the circuit's
location without emitting the op.