[HWMemSimImpl] Randomize each memory word separately (#2860)
Instead of assembling a randomized word value in a separate variable and
then storing that to every memory location, this commit extends
`HWMemSimImpl` to generate new random values on-the-fly for every memory
location as it iterates over them. As a result the initialization code
becomes shorter and the memory is fully randomized.