[FIRRTL] Fix instance port attribute erasure bug
Fix a bug in RemoveUnusedPorts where when a new instance was created,
this would not copy over the "output_file" attribute.  This created an
interaction between Grand Central (which sets the "output_file" on
instances) and RemoveUnusedPorts (which could previously delete the
"output_file" on instances).  This bug would manifest as invalid XMRs
because these instances would be extracted to a generic "bindfile" and
not to the expected Grand Central output bindfile (which a build flow
would likely not load).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>