[FIRParser] Default chisel3 asserts to `ifElseFatal` (#2876)
In SFC, these `chisel3_builtin` asserts default to the `ifElseFatal`
format.  We previously thought that we could emit these in a simpler
format, but it is missing the wrapper for `ASSERT_VERBOSE_CONDITION`.
This macro is used to prevent assertions from triggering when the test
bench is reseting.  Without guarding the assertion on this value, we are
getting some spurious failures before the first reset has run.