[FIRParser] Fix partial connect expansion with analog in bundles (#2925)
When the LHS and RHS of a partial connect were the same types, we would
just emit a regular connect.  We forgot to check if there were any
analog types inside the bundle When this is the case, we have to recurse
on the bundle elements and attach the analog fields together.