[Support] Make BackedgeBuilder use registered ops (#2927)
Previously, BackedgeBuilder was using an unregistered dummy op. This
required setting the context to allow unregistered ops, which
BackedgeBuilder did but that wasn't thread-safe. (An upstream update
asserts when one tries to set that in the presence of threads.)

This PR has BackedgeBuilder use a registered op, which users can select
via a template parameter. It also adds a standard `BackedgeOp` to the
`hw` dialect which should cover the vast majority of uses.