[StandardToHandshake] enable reuse of region lowering (#2986)
This change makes the lowering infrastructure used to lower `func::FuncOp` bodies available for other usages. As discussed in #2951 the lowering of the region is not dependent on the operation holding it.

The newly added `lowerRegion` function expects a sub-type of `HandshakeLowering` which might have overridden member functions. The members are declared `virtual` to allow for such overriding without re-implementing the `lowerRegion` function.