[FIRRTL][InferWidths] Support bundles in mux constraint (#3071)
The infer widths pass made the assumption that the arguments to a
`MuxPrimOp` were ground types, and would trigger an assertion trying to
create a `max(high, low)` constraint on the operands.  This changes the
mux constraint generation to recursively constrain the type of the
result to the maximum of the high and low signals.