[FIRRTL] Composable LowerAnnotations (#3129)
* [FIRRTL] Remove all-or-nothing LowerAnnos

Change the behavior of the LowerAnnos pass to work with existing
annotation scattering.  Previously, use of the `--new-anno` firtool
flag (and backing FIRParser option) would cause ONLY the LowerAnnos pass
to be used.  This changes that to use normal scattering and then apply
LowerAnnos.  This is done to enable piecemeal migration of scattering
work into LowerAnnos.

* [FIRRTL] Make LowerAnnos search non-standard attr

Change the LowerAnnotations pass to only process annotations from a
non-standard attribute on the circuit with key
"annotationsForLowerAnnotations".  This is done to enable the
LowerAnnotations pass to be run _after_ normal annotation processing
inside FIRAnnotations and FIRParser.  This then enables us to migrate
annotations one-by-one from FIRAnnotation/FIRParser handling (by moving
the annotation into "annotationsForLowerAnnotations").  This commit does
not put any annotations into this area for processing.  That will be
handled by subsequent commits.

* [FIRRTL] Add LowerAnnos passthrough in Parser

Modify the FIRRTL parser to let annotations that come in with the key
"annotationsForLowerAnnotations" to be placed in an attribute on the
circuit with this key.  This is currently unused infrastructure that
lets annotations inside FIRAnnotations be put behind this key to let
them be handled by the LowerAnnos pass.

* [FIRRTL] Add isAnnoClassLowered util, NFC

Add a utility, isAnnoClassLowered, that returns true if an annotation
class name is known to be handled by the LowerAnnotations pass.  This
effectively leaks internal information from the LowerAnnotations-defined
table of known annotations to let parsing/scattering logic know that
these annotations will be handled later.

* [FIRRTL] Skip LowerAnnos Annos During Parsing

Modify the Annotation parser to NOT handle annotations that are known to
LowerAnnotations.  This is done to enable seamless migration from
annotation handling in the parser to annotation handling in
LowerAnnotations.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>