[FIRRTL] Move tryGetAs util into anno header, NFC
Move the tryGetAs utility from a static method inside FIRAnnotations
into the common FIRRTLAnnotationHelper.h header.  This utility has broad
application for generating a standard error message if an annotation
fails to have a specific field of a specific type.  Narrowly, this is
done to make this available for Grand Central annotation handling which
will live inside the Grand Central pass in later commits.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>