[FIRRTL] Make DontTouchAnnotation an Anno, not Sym
Change DontTouchAnnotation from a symbol back to an actual annotation.
This is done because we need to have a true notion of "optimization
barrier" which DontTouchAnnotation implies.  However, symbols are used
almost everywhere to provide anchors for non-local annotations.  In
effect, this relaxes the explosion of "optimization barrier" behavior
which is functionally correct, but results in much less optimized
Verilog for FIRRTL code.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>