[FIRRTL] Make namespace constructible from ports/body ops
Add a `addPorts` and `addBody` method to the FIRRTL `ModuleNamespace`,
allowing it to be constructed from just a plain list of ports or the
ports of a module, or just from a module's body ops. This allows for
more specialized usecases where only name uniquification on the port
list is required.

Signed-off-by: Fabian Schuiki <fabian@schuiki.ch>