[PyCDE] Specialize `Instance`s (#3160)
When creating `Instance`s, check the static op and instantiate a
specialized class. Currently, two specializations exist:
`ModuleInstance` and `RegInstance`. The former is the only container
type for now, so move anything related to children there.

Most of this is refactoring to support specialization. There is very
little new code.