[Support] Add iterator interface to SymbolCache (#3142)
Continuing from the work in #3137, instead of introducing a new class for generating unique names, it might be better to simply have a way to initialize a namespace from a SymbolCache.

To do so, this PR introduces an iterator interface to the base SymbolCache class which iterates over pairs of `{mlir::Attribute, mlir::Operation*}`. My experience with iterators are limited, so this might be the wrong way to go about creating an iterator that is accessed through a virtual base - please let me know.

Given this interface, it's now possible to range-loop over any given cache and use the values for namespace initialization.
I've included changes to `Namespace.h` and `HWSpecialize.cpp` in this commit to prove that things work.
If this PR is something that we find useful, then i'll factor those changes out before a merge.