Temporarily stop printing version info in generated files (#3169)
In 91ea7871c we started adding a line to print CIRCT version information
in all generated files.  Since this code did not discriminate which
files should have the string, it would create illegal content for some
file formats.  For now, this just reverts the version printing code.  In
the future, we should only emit this string in Verilog files.