[FIRRTL] GCT Data/Mem Tap Scattering in LowerAnnotations (#3186)
Move all logic related to Grand Central Data and Memory Tap annotation
scattering from `FIRAnnotations.cpp` into the `LowerAnnotations` pass.

Besides moving the scattering into `LowerAnnotations`, there are a few
additional changes to the actual annotations that are being generated.

Data and memory tap annotations identify an extmodule which is to be
synthesized into a data/mem tap module by the GCT pass. To capture which
synthesized port should tap which signal in the design, annotations are
generated in pairs: one on the extmodule port and one on the thing being
tapped. These used to share the same class name and were only
distinguishable by a `type = "portName"` and `type = "source"` field.
This commit removes this `type` field and generates annotations with the
distinct classes in the following cases:

- `ReferenceDataTapKey` becomes `ReferenceDataTapKey.port` and
  `ReferenceDataTapKey.source`, respectively.
- `DataTapModuleSignalKey` becomes `DataTapModuleSignalKey.port` and
  `DataTapModuleSignalKey.source`, respectively.
- `MemTapAnnotation` becomes `MemTapAnnotation.port`, `MemTapAnnotation.source`,
  and `MemTapAnnotation.blackbox` respectively.
- `DataTapsAnnotation.blackbox` is added.

Furthermore, the `DataTapsAnnotation` that used to be placed on the data
tap extmodule is now called `DataTapsAnnotation.blackbox`. This change
is necessary to since `LowerAnnotations` has to distinguish between the
`DataTapsAnnotation` that requires custom scattering and the subsequent
simplified `DataTapsAnnotation` generated by that scattering which is
simply to be placed on the data tap extmodule.

Integration tests required no modifications, which is reassuring since
this commit attempts to only modify the intermediate representation of
GCT annotations before they are handled in the GCT data taps pass. This
GCT data taps pass tests were updated to match the new annotations. The
annotation scattering tests were migrated from `annotations.fir` into
`annotations.mlir` to match the move from `FIRAnnotations.cpp` into
`LowerAnnotations`.

This commit also adds a test for GCT mem taps on an actual `firrtl.mem`
op again which seems to have gotten lost sometime in the past and caused
a failure in an integration test that wasn't caught in the GCT data taps
pass tests.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
Signed-off-by: Fabian Schuiki <fabian@schuiki.ch>
Co-authored-by: Fabian Schuiki <fabian@schuiki.ch>