[FIRParser] Fix bug extending an APInt to a smaller width (#3226)
We recently removed usages of the deprecated function `sextOrSelf`.
This function allowed you to extend an APInt to a smaller width than it
currently is, becoming a noop.  The replacement function does not allow
doing this, and asserts.  During the cleanup I missed that we had to
guard this usage of `sext` to check the initial bit width before
attempting to sign-extend.

See: https://github.com/llvm/circt/pull/3202