[FIRRTL] Update EmitOMIR for New NLA Style (#3336)
Remove internal NLA breadcrumbs that are unnecessarily generated in
EmitOMIR.  These are entirely unused by internal EmitOMIR analyses and
serve no purpose other than to be removed by the pass later.

Change the EmitOMIR pass to remove no hierpath ops.  It cannot safely do
this without computing the uses of a hierpath.  While the pass could be
updated to do this, it would be cleaner to handle hierpath removal with
a common DCE pass.

This change is made to support new NLA-style paths.  With this new
style, multiple NLAs may share the same hierpath ending at a module and
it is thereby unsafe for NLAs to decide to delete a hierpath without
checking for users.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>