[FIRRTL] Block Removal of Components w/ Annos
Make the presence of an Annotation block the removal of a component.
The presence of an Annotation indicates that some later pass cares about
this wire and it cannot be safely deleted.  Concretely, this enables
downstream passes from NOT having to slap DontTouchAnnotations on things
they care about (which implies, much strong, optimization block
behavior).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>