[PyCDE] Add NDArray class (#3369)
A PyCDE NDArrayserves as a Numpy view of a multidimensional CIRCT array (ArrayType).
The Matrix ensures that all assignments to itself have been properly converted to conform with insertion into the numpy array (circt_to_arr).
Once filled, a user can treat the NDArray as a numpy array. The underlying CIRCT array is not materialized until `to_circt` is called.

The NDArraycan be used both directly, as well as through PyCDE `ListValue`s. Users who use the `NDArray` directly have access to all numpy functions.
In reality, it will only be a subset of the numpy array functions which are safe to be used in a PyCDE context. Since using the numpy transformations is experimental, curating the set of allowed functions at the level of `ListValue`s seems like a safe starting point.