[FIRRTL] rework InnerRef verification, starting with HierPathOp (#3366)
InnerRef's are expensive to verify presently, as the InnerSymbol
portion requires walking the target operation (FModuleLike).

HierPathOps are chains of these, and so were especially
problematic in this regard.

This is a similar problem to verifying Symbol (SymbolTable) users,
and is solved in a similar way.

The primary components of the new approach:

1. Drive verification from the top of InnerRef resolution scope

Only walk the relevant IR once, parallelizing where makes sense.

2. Introduce an interface for verifying InnerRef's

Operations using InnerRef's can implement the interface
to opt-in to this new verification.

For (1), a new trait "InnerRefNamespace" is added to mark Operations
in which InnerRef's are resolved.  For now this is only added to CircuitOp.

For (2), introduce "InnerRefUserOpInterface" with a "verifyInnerRefs"
method, similar to SymbolOpUserInterface and verifySymbolUses.
Operations using InnerRef's should prefer this interface, at least for
purposes of verification: since `verifyInnerRefs` also provides the operation
with a SymbolTable, all the same verification done in `verifySymbolUses` can be
done together with verification involving resolving the InnerSymbol's.
(There's no need for a SymbolTableCollection within an InnerRefNamespace,
as there's only one SymbolTable, at the top level.)

In addition, InnerSymbolTable is fleshed out a bit so that it can be
used for basic lookups of InnerSymbols, and "InnerSymbolTableCollection"
is added that serves a similar purpose as SymbolTableCollection.
InnerRefNamespace is also a class (as well as the trait) that combines a
SymbolTable and the InnerSymbolTableCollection, for convenient access to both
in "verifyInnerRefs".

Some structural requirements are now explicitly checked.