[FIRRTL] EmitOMIR: Drop temporary HierPath's that may be invalid (#3384)
This pass creates temporary NLA operations which are known to be dead
and sometimes are invalid, causing breakage during subsequent circuit
verification. The invalid NLA encountered is when the target is in the
root module, and so we end up making a one-element (InnerRef) NLA.

While we cannot delete incoming HierPath's attached to operations in
input IR without a more careful analysis of all possible users, any NLA
explicitly created in this pass can be reasoned about and removed at the
end for cleaner (and valid) IR.

This change restore behavior for deleting the HierPath's added in this
pass which was accidentally removed.

This fixes assertion failure on a real design.