[FIRRTL] rm GCT-SM Invalid Source/Sink Test, NFC
Remove a test from Grand Central (GCT) Signal Mappings (SM) that was
checking that an illegal source/sink that did not exist in the circuit
was dropped from the output.  This should really never be happening and
we should be erroring if this ever occurs.

This commit is part of staging larger changes to migrate GCT-SM
annotation handling from parsing to the LowerAnnotations pass.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>