[FIRRTL] Use LowerAnnotations in firtool test, NFC
Modify firtool tests to use Annotations which are known to
LowerAnnotations as opposed to using non-standard Annotations without a
class field.  This makes LowerAnnotations load-bearing for these firtool
tests and also helps to differentiate that firtool does _not_ scatter
annotations.  This is done as incremental work towards moving all
Annotation handling to LowerAnnotations.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>