[FIRRTL] Improve Error on Instance Target Subindex
Improve the error message when trying to resolve a Target on an instance
that uses a subindex.  E.g., this makes the error for "~A|A>b[0]" better
by indicating that this is an illegal index of an instance.  Previously
the message would report about being unable to find port "0".

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>