[FIRRTL] Move Anno tests to LowerAnnotations, NFC
Migrate all existing Annotation tests that are relying on FIRRTL Parser
logic to use LowerAnnotations instead.  This is a direct conversion of
the tests from ".fir" to ".mlir".  The only change in behavior in the
tests is related to instance port annotations.  LowerAnnotations always
moves these to the module whereas the FIRRTL Parser would leave these on
the instance.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>