[HW] [ESI] Add module mutation interface, use it in ESI pass (#3532)
Adds a new `HWMutableModuleLike` op interface which exposes a bunch of methods which were all present in the `hw` operations which implemented `HWModuleLike`. This allows one of the ESI passes to not have knowledge of any `hw` module or instance ops -- it can operate on anything which has a module implementing `HWMutableModuleLike`. Paves the way to supporting `msft.module`.