[PyCDE] Add shorthand FSM builder function (#3560)
Generate a FSM from a dictionary of states and their transitions.

E.g.:
```
{
    "a": [
        ("c", "go"),
        "b",
    ],
    "b": [],
    "c": []
}
```

creates an FSM with 3 states (a, b, c). 'b' and 'c' have no outgoing transitions, and 'a' has two outgoing transitions. The first transition ("c", "go") transitions to 'c' whenever an input port 'go' is asserted. The second transition is a default transition to 'b'.

Any state and guard referenced within the dictionary will automatically be created as a state operation and top-level input, respectively.