[SystemC] Add SignalOp (#3614)
Limits the allowed types for signals to signless integers for now. We can expand that later to all the possible types allowed according to IEEE 1666-2011 §6.4.3. as needed.
Aims to implement names not as hints but guarantees that the name given to a signal, port, etc. during lowering to SystemC is preserved up to emission.