[LowerToHW] Delete attach op if there is single source (#3627)
This PR implements an optimization to delete attach op if there is only a
single source value (usually port or wire). The optimization happens in the 
lowering of instance ops. If an instance result has an analog type and its user
is only attach op, then we try eliminating the attach op. If there is a single 
source value in the attached values, we use the value as a lowering result 
of instance port result.