[FIRRTL][LowerToHW] Fix issue with backedges not being replaced (#3619)
In LowerToHW we have a hashtable which maps every FIRRTL value to its
lowered HW value. Some "no-op" casts, such as `asClock`, are removed
during lowering, so their result value is actually mapped to its
operand's lowered value.

For instance input ports, we lower them using backedges to take
advantage of the graph structure of the HW dialect and eliminate some
extra wires. The port's value is mapped to the backedge, until we find
the connect operation driving it, at which point the backedge is
replaced with its true value, and the lowering table is updated.

We can have a situation where an input port is lowered to point to a
backedge, and a no-op cast of the port is also lowered to point to the
backedge. When the backedge is replaced, the lowering table is only
updated for the input port and not the cast operation. This means that
any further uses of the cast's result will lower to uses of the
backedge, which are never fixed up.

This change delays the replacement of the backedge with its value until
all operations have been lowered. This means that uses of the cast
which lower to uses of the backedge will be properly replaced.

This commit also addresses a related issue, when an input port is
connected to another input port.  The backedges from one port need to
"see through" the backedge that is driving it to find the real value
driving the net.