[FIRRTL] Use annotations to represent dontTouch (#3639)
This commit switches back the representation of dontTouch to normal
annotations. We have used symbols to uniformly represent optimization
blocking behavior (#2407), but it has become hard because of the complexity
hierpath updates or field-sensitive symbols. Therefore, this commit
changes the representation of dontTouch to normal annotations.

First, LowerTypes is modified to emit errors for aggregate symbols.
This assumption makes LowerTypes a lot simpler. Actually, symbols are
attached before LowerTypes only by dontTouch so the restriction should
not be a big problem. Second, since we don't have to attach symbols at
LowerTypes, the implementation of `filterAnnotation` is simplified a lot.
Also, the renaming of symbols will not happen anymore so they are 
removed as well.