[FIRRTL][SFCCompat] Add support for aggregates (#3670)
This adds support for aggregate typed registers to the SFC compat pass.
SFC Compat looks at the values which are connected to registers; if they
are all invalid the reset can be removed, and resets for async registers
require a constant reset value.

Previously, this used the function `getModuleScopedDriver` which only
handles ground types.  This creates a function `walkDrivers` which
performs a similar operation, but applies a call back when it finds
which fields of the aggregate is driven by some other value. Since the 
source value and the the destination value might both be aggregates, the 
callback uses FieldRefs to indicate the field.