[Support] CPPCDE - C++-based CIRCT design entry
CPPCDE intends to reduce the verboseness of emitting large amounts of CIRCT IR by providing in-CIRCT generators.

One example is the IR generated in the `HandshakeToFIRRTL` pass - locally, I'm able to reproduce operator implementations to 33% their `HandshakeToFIRRTL` size using this. As an added incentive, this kind of in-CIRCT generator infrastructure has been a long-standing prerequisite for me to start working on getting `HandshakeToHW` finished.

Is this a fully-fledged DSL? ...no - the intentions are the same as with PyCDE, that is, to provide a layer of syntactic sugar on top of the builder infrastructure to facilitate writing IR generators. A further restriction is that i intend (at least to begin with) for this to be strictly a library that other passes can leverage.

The implementation shares similarities to PyCDE, and takes a dynamic approach whereever possible. A more static, typesafe, approach would have been to e.g. create ports as member variables. The intention here is to allow a more flexible generator infrastructure - generators of generators - which we've seen useful in PyCDE. Also, this implementation is simpler and more like 'syntactic sugar' instead of a full-blown DSL.

Error handling is currently done by asserts - not ideal. Exception raising would be preferred (caught in the `go` method, which would return a `LogicalResult`) but currently not possible, due to the project being built with exceptions disabled.