[COMB] Add flag to assert binary behavior.
Comb exists in a place where it might or might not need to deal with Verilog or VHDL weirdness.  Some transformations which are desirable might not be semantic preserving with respect to concepts such as 'x'.  This is overly strong for a number of languages.  An "bin" flag is added to most operations to allow flagging the comb operation as only requiring binary semantics.  Frontends by construction, or by analysis can enable this flag to enable richer comb-level optimizations.