[FIRRTL] Backport IMDCE updates to sifive 1.5 (#3771)
* [FIRRTL] Remove empty modules in IMDCE (#3378)

This commit extends IMDCE to delete empty modules. A module is empty if
it doesn't have any argument, operation and annotation. In the post-processing,
we traverse modules in a post-order of the instance graph, and remove empty
modules and their instances in a bottom-up manner.

* [IMDCE] Forward constant output ports to caller sides (#3688)

This PR makes IMDCE propagate constant output ports to caller sides before
actually performing dataflow analysis so that we can eliminate constant output ports.