[FIRRTL] ModuleInliner: Add support for RefType (#3739)
This PR adds support to handle inlining of RefType ops. Temporary wires of
RefType that are created to replace ports of inlined modules, must be removed.
As a post-processing step, the pass iterates over all the `RefType` wires, which
were added during the inlining transformation. Then replaces all the uses of the
wire with its driver, and marks the wire for removal. Some of the uses of the
wire need to be moved to ensure the driver dominates all the uses.