[firtool] Reorganize pass controlling options (#3794)
We have many flags which control whether or not we run a specific pass,
such as `--lower-types`.  Most of these flags should never be used by
regular firtool users; they are only really useful for circumventing a
crash in a pass and can often lead to other crashes.  This moves most of
the pass options to be of the form `--disable-mypass`, and makes them
 hidden by default.  The descriptions of these options were made
uniform, at the cost of a less detailed description.

The `GrandCentral` pass was enabled by default. Since its behavior is
controlled by annotations, it should always run.

This also did a small sweep to make sure that all our option
descriptions begin with a capital letter.