[FIRRTL] Support GC internalPath on Modules (#3829)
In GrandCentral DataTapsAnnotations, there are DataTapModuleSignalKeys
which tap signals based on a string `internalPath`. These are typically
used on extmodules with the promise that the `internalPath` exists
somewhere in the implementation of the extmodule. Generators that cannot
distinguish between they are constructing Grand Central annotations on
modules or extmodules in their emitted FIRRTL use these unsafe
`internalPath`s on actual modules.

This is a patch fix for something we can better implement once
GrandCentral is migrated to using RefOps.