[FIRRTL][GrandCentral] Generate verbatim based on RefType Values (#3852)
This commit generates verbatim embedded cross-module-reference based on
`RefResolveOp`. If the `AugmentedGroundType` annotation can be traced to a
`RefType` value, then just use the `verbatim` `Value` substitution instead of
the symbols substitution. This updates the indices used in the verbatim text,
 for the value starting at 0 and following indices for symbols.
Added lit tests, for the `RefType` value, constant value and symbols.