[SystemC] Add VariableOp and AssignOp (#3830)
Add an operation to model C++ level variable declarations and one to assign to them. This also includes the emission patterns for ExportSystemC. Note that the variable op implements custom parser, printer, and verifier because the single optional operand makes it difficult to use the predefined constraints to check for type equality and infer the type from the result type.