Bump LLVM (#3872)
* [Calyx][FIRRTL] Put explicit type hints for TypedValue where necessary.

With the introduction of TypedValue in upstream llvm commit
688c51a5acc53b456014e53663051476d825e896, there are now a few cases
where we need to explicitly write the types. In some cases, we need to
explicitly upcast to Value to be consistent with other Values in the
expression, and in other cases, we need to upcast the type parameter of
TypedValue when there are ambiguous implicit conversions.

* [FIRRTL] Replace llvm::GreatestCommonDivisor64 with std::gcd.

GreatestCommonDivisor64 was removed in upstream LLVM in
87c38323a2cff5b26023b24c36a7c01741aba834, and the recommended migration
path is to use std::gcd. This performs an explicit cast to uint64_t for
the same reasons described in 4a2377afd69bcf014492cb665ee955eab3121c4c,
which is that std::gcd may have different and unexpected semantics when
its arguments are of different types and signedness.

* [NFC] Explicitly set dialect accessor prefix to kEmitAccessorPrefix_Raw.

Upstream recently changed the default value of emitAccessorPrefix to
kEmitAccessorPrefix_Prefixed. This patch explicitly sets
emitAccessorPrefix to the previous default value of
kEmitAccessorPrefix_Raw, which will allow us to the new accessor prefix
style more gradually and at our own pace. Upstream will completely
remove the _Raw style in the next couple of months, so we will need to
make the migration soon, though.

See [1] for more information.

[1]: https://discourse.llvm.org/t/psa-ods-generated-accessors-will-change-to-have-a-get-prefix-update-you-apis/4476

* [firtool] Pass config object to new writeBytecodeToFile API.

The third argument to writeBytecodeToFile() now takes a
BytecodeWriterConfig object instead of just the producer string.

* Bump llvm to 13f1bc41888e7d6555c532ba5fa925b9fe3e6b2f.