[FIRRTL] Check each bundle field to infer memory port kind (#3906)
Update the logic to get the memory port kind. Match bundle field names to get the PortKind. 
If fields are `addr`, `en`, `clk`, `data`, then it is a Read port
If fields are `addr`, `en`, `clk`, `data`, `mask` then it is a Write port
If fields are `addr`, `en`, `clk`, `wdata`, `wmask`, `rdata`, `wmode` then it is a ReadWrite port
This matches the field names exactly, instead of just relying on the field count.