[FIRRTL] Fix Annotation Parsing of String Integers
Fix a bug in how Annotation JSON Parsing worked where a string
number (e.g., "0") would be parsed as a number (e.g., 0).  This was a
source of bugs in downstream Annotation handling where an Annotation
would expect that something was a string when it would instead show up
as a number.  Concretely, this was creating issues in Grand Central View
Annotation handling where a field of an AugmentedBundleType would show
up as "0 : i64" instead of a string.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>