[ESI][Services][NFC] Adding an OpInterface for declarations and making arg/result names more consistent (#3927)
Added the `ServiceDeclOpInterface` which is intended to be implemented by any op which'll be a service declaration. Has methods to get the list of service ports and validate a connection request. The introduction of a struct which describes a service port allowed me to clean up a bunch of code. This changed the error messages, requiring an update to the errors test.

Also took this opportunity to standardize terminology on `toClient` and `toServer` to provide better clarity. Was helpful in debugging this change.