[firtool] Lower annotations with -parse-only
Change `firtool -parse-only` to stop after parsing, verification, and
annotation lowering as opposed to stopping after parsing.  It is
incredibly rare that a user (or, much more likely, a developer) actually
wants the behavior of stopping after parsing as opposed to stopping
after annotation lowering.  This avoids a frequent pattern where a
user/developer will use something like the following instead of
`-parse-only` for any circuit that includes annotations:

  firtool -parse-only Foo.fir | circt-opt -firrtl-lower-annotations

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>