[FIRRTL] Allow foreign types in module/instance ports (#2694)
In order to support partial lowerings involving the FIRRTL dialect, and
to mix in foreign operations and types, we need a way to pass foreign
types through FIRRTL modules, instances, and trivial connects.

This commit extends `FModuleLike`, `InstanceOp`, `StrictConnectOp`, and
`WireOp` to also accept foreign, non-FIRRTL types for ports and
connections to those ports. It also adapts the `LowerToHW` pass to
simply pass through foreign-typed values when they appear in module and
instance ports. This requires adding special case handling in a few
places since we cannot create an `InOutType` and therefore `sv::WireOp`
around arbitrary foreign types, so we generally have to immediately
materialize foreign values.

This lowering can be quite usefully combined with the builtin
`unrealized_conversion_cast` to wrap and unwrap values of foreign types
and establish a path for FIRRTL values to cross dialect boundaries in a
controlled fashion.

In more details, this PR does the following:

- Allow foreign types in:
  - `FModuleLike`
  - `InstanceOp`
  - `StrictConnectOp`
  - `WireOp`
- During lowering to HW, force module and instance port connections to
  materialize connected values directly if they are of foreign type,
  since we cannot create temporary wires to hold these values.