[PrepareForEmission] Improve namehints heuristic and add mux heuristic (#4019)
This PR improves namehints heuristic and add a heuristic to spill wires for mux. Namehint heuristic is modified to spill wires for namehints that doesn't have "\_" prefix regardless of term sizes. For expressions with "\_" prefix, if the term is greater than a threshold, the expression is spilled to a wire. Also this PR adds a heuristic option to spill all mux.

Co-authored-by: Will Dietz <will.dietz@sifive.com>