[Seq] Add const-clock and const-reset folders for FirRegOp (#4090)
Introduce additional folders and canonicalizers for the `FirRegOp` that
deal with registers that have a constant clock or reset value. The
FIRRTL spec states that the value of a register before the first clock
edge is undefined (unless reset through an async reset), which allows us
to pick a convenient value in the cases where a register is provably
never clocked or never reset.

This also moves the canonicalizer for self-referential registers into
the folder since it does not need to create arbitrary operations.