[LLHD] Update signal buffer loading to use LoadIntFromMemory helper. (#4102)
This code is tripping ASAN warnings. Taking a suggestion from @youngar
in the linked issue, it seems like there is a helper in LLVM to do
what we want here. If I drop that suggestion in, the tests still pass
and the ASAN warnings are fixed. I haven't thought about it deeply,
but this seems to make sense. The comment suggested some concerns
about handling big endian systems, which LoadIntFromMemory appears to
already take into account.

Closes https://github.com/llvm/circt/issues/1071.