[PyCDE] Fix sporadically failing integration test
If the RTL simulator doesn't compute a result in time, the esi_test.py
fails. This is actually a tough one to fix properly since the test
checks to see that there is not a result on any of the other channels. I
think the best way to fix this is to just wait a small amount of time.