[SystemC] Add interop.verilated operation (#4104)
This operation instantiates a hw module that is or will be verilated and
allows to generate the necessary interoperability layer to talk to the
verilated module in various contexts, i.e., in a SystemC module, a
Verilog module, inside a CIRCT based simulator, etc.
Having a separate instance operation avoids transformations performed on
the regular hw instance such as inlining, etc. and allows to add more
attributes to specify the configuration under which the instantiated
module was verilated.