[Comb][HW] Fold tree of muxes into an array_get
If a tree of muxes uses N bits to select from 2^N elements, it is replaced with an `array_get` and `aarray_create`.

Presently the index is concatenated from arbitrary bits. This can be replaced with a pattern identifying a bit select from a number and the transformation would still be useful.