[FIRRTL] Add port name to InferResets error
Include the port name when generating an InferResets uninferred port
error.  This is the bare minimum to help debug these errors as opposed
to having to rely on the user to guess which port failed.

Port source locators are not available from Chisel (nor do we store them
in CIRCT if they did exist!), so this is the best we can do for now.

Fixes #4183.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>