[SV] Add ReadMemOp to model $readmemb, $readmemh
Add a new operation to the SystemVerilog Dialect: ReadMemOp.  This
operation exists to model both "$readmemb" and "$readmemh".  These
operations cause a specific memory to be loaded with the contents of a
file.

These operations do not (yet) model the ability of their Verilog
counterparts to load part of a memory.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>