[ExportVerilog] Fix i0 value emission of struct operations (#4208)
This PR fixes invalid emission of i0 values contained in structs. 
Currently, i0 values in struct types are emitted as `logic` but `logic` is a 1bit width 
value so it causes width miss matches in many cases. This fixes that issue by just 
eliding the field entirely and emitting them as comments.