[FIRRTL] Fix GCT View Directory Behavior for DUT
Fix the behavior of Grand Central (GCT) Views directory placement for
modules which are instantiated in the GCT View Companion module and also
in the testbench, but not the DUT (as indicated by a MarkDUTAnnotation).
Previously, the check would only move modules exclusively instantiated
under the companion.  This would cause modules instantiated in the
testbench (above the DUT) to wind up in the DUT directory and cause
problems for downstream physical design tooling that wants to only see
design Verilog in the DUT directory.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>