[ExportVerilog] Don't emit a constant if it has zero width (#4239)
When emitting a module instance which has zero-width ports, don't try to emit the zero-width expression if it's a constant since that results in an error and some ugly lookin' output (though the resulting verilog works since the whole port connection is commented out). In cases other than `hw.constant`, valuable information can still be printed so we'll keep this as a special case (for now).