[FIRRTL][Dedup] Fix DontTouch annotation handling (#4241)
When we dedup modules, we combine the annotations from both. Since
DontTouch is a sticky annotation and there is no point in having more
than one, we tried to only keep one copy of the `DontTouch` annotation.
This logic did not take into account that the annotation might be
targeting a specific field of an aggregate, and could end up
deduplicating the `DontTouch` annotations which targeted different
fields.

This PR uses a set of annotations to take into account the `fieldID` of
the annotations.  It also started ripping out the `circt.nonlocal` field
from any `DontTouch` annotation, which has no effect on the annotation,
and just hurts our ability to deduplicate these annotations.