[FIRRTL] Add WiringProblem to LowerAnnotations
Add WiringProblem infrastructure and enable its use in
LowerAnnotations.  This defines an array of WiringProblems that a user
can add to during LowerAnnotations.  Each WiringProblem is a request for
a source Value to be connected to a sink Value in the design using
RefType ports.  LowerAnnotations is then changed to solve all
WiringProblems are bore RefType bores through the design.

This is done in preparation for switching Grand Central Taps and Views
to using this infrastructure instead of its current poorly performing
algorithm that bores ports one at a time.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>