[FIRRTL] Block dedup on input RefType ports
Change FIRRTL's structural deduplication ("dedup") pass to block
dedup if a module has input RefType ports.  This is done because input
RefType ports are consumed by a RefResolveOp which will lower to a
Verilog hierarchical reference (XMR).  If dedup is allowed, this XMR may
not have a valid representation because it refers to two different
things in the circuit.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>