[FIRRTL] Add Inner Symbol DCE Pass (#3120)
This finds and removes any InnerSymAttrs that are not used.

It starts with a walk through the IR to find any InnerRefAttrs. This
looks through all operations' attributes, descending through nested
attributes using the sub element interface.

After finding all live InnerRefAttrs, it walks through each module in
parallel looking for operations with an InnerSymAttr. If the module
and symbol were not in the set of live inner references, the inner sym
is removed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
Co-authored-by: Mike Urbach <mikeurbach@gmail.com>