[PyCDE] PyTorch (dot product) integration test (#4352)
Introduces an integration test which takes a PyTorch kernel, builds an ESI system around it, then executes a cosim test. We commit and use MLIR assembly as as to avoid (1) having to install torch_mlir, (2) having to register the cf/arith/memref dialects, and (3) changes to torch_mlir or upstream dialects breaking this test.