[ExportVerilog] Sink mux inlining check to isExpressionEmittedInline. (#4392)
Before, PrepareForEmission would check the disallowMuxInlining option
before checking isExpressionEmittedInline. In this case, we might
choose to spill muxes into generated wires that are themselves
trivially assigned to an output, wire, or reg.

Instead, this change sinks the disallowMuxInlining check into
isExpressionEmittedInline, after we check if the expression is part of
an output or assignment. This way, muxes are allowed to be inlined
directly into an assignment we are going to generate anyway, avoiding
unnecessary generated wires. Muxes will still be spilled to wires
otherwise, when requested.